fix: fix imports and finishing componentization
