+++ lib/wallet/view/wallet_screen.dart
@@ -1,8 +1,8 @@
-import '../widgets/wallet_body.dart';
 import 'package:flutter/material.dart';
 import 'package:hooks_riverpod/hooks_riverpod.dart';
 
 import '../../shared/widgets/default_navbar.dart';
+import '../widgets/wallet_body.dart';
 
 class WalletScreen extends StatefulHookConsumerWidget {
   const WalletScreen({Key? key}) : super(key: key);
